refactor: package modules under jobscraps
