+++ jobscraps/__init__.py
@@ -1,3 +1,15 @@
 """JobScraps package."""
 
-__all__ = []
+from .scraper import JobScraper
+from .duplicate_manager import DuplicateManager
+from .database import DatabaseConfig, JobDatabase, BackupMixin
+from .config import JobSearchConfig
+
+__all__ = [
+    "JobScraper",
+    "DuplicateManager",
+    "DatabaseConfig",
+    "JobDatabase",
+    "BackupMixin",
+    "JobSearchConfig",
+]
