+++ scripts/db_stats_cli.py
@@ -457,10 +457,12 @@ def search_jobs(
             console.print(f"[bold red]Database error: {str(e)}[/bold red]")
 
 @app.command("run")
-def run_scraper(script_path: Optional[str] = typer.Option(None, help="Path to scraper.py")):
+def run_scraper(script_path: Optional[str] = typer.Option(None, help="Path to jobscraps/scraper.py")):
     """Run the job scraper script"""
     if script_path is None:
-        script_path = os.path.join(os.path.dirname(os.path.dirname(__file__)), "scraper.py")
+        script_path = os.path.join(
+            os.path.dirname(os.path.dirname(__file__)), "jobscraps", "scraper.py"
+        )
     
     if not os.path.exists(script_path):
         console.print(f"[bold red]Scraper script not found at {script_path}[/bold red]")
